Moving BananaAssertions
